api/requirements.txt
fastapi
supabase
openai
pydantic
python-dotenv
httpx
postgrest
